+++ R/expect-output.R
@@ -7,6 +7,25 @@
 #' Note that warnings are captured by a custom signal handler: this means
 #' that `options(warn)` has no effect.
 #'
+#' @section Testing `message` vs `class`:
+#' When checking that code generates an error, it's important to check that the
+#' error is the one you expect. There are two ways to do this. The first
+#' way is the simplest: you just provide a `regexp` that match some fragment
+#' of the error message. This is easy, but fragile, because the test will
+#' fail if the error message changes (even if its the same error).
+#'
+#' A more robust way is to test for the class of the error, if it has one.
+#' You can learn more about custom conditions at
+#' <https://adv-r.hadley.nz/conditions.html#custom-conditions>, but in
+#' short, errors are S3 classes and you can generate a custom class and check
+#' for it using `class` instead of `regexp`. Because this is a more reliable
+#' check, you `expect_error()` will warn if the error has a custom class but
+#' you are testing the message. Eliminate the warning by using `class` instead
+#' of `regexp`.
+#'
+#' If you are using `expect_error()` to check that an error message is
+#' formatted in such a way that it makes sense to a human, we now recommend
+#' using [verify_output()] instead.
 #' @inheritParams expect_that
 #' @inheritDotParams expect_match -object -regexp -info -label
 #' @param regexp regular expression to test against.
@@ -78,10 +97,12 @@ expect_error <- function(object,
 
   if (!is.null(act$cap)) {
     if (!simple_error(act$cap) && is.null(class) && !is.null(regexp)) {
-      klass <- paste0(class(act$cap), collapse = "/")
+      klass <- paste0(encodeString(class(act$cap), quote = '"'), collapse = ", ")
       warn(paste0(
-        act$lab, " generated a condition with class ", klass, ".\n",
-        "It is less fragile to test custom conditions with `class`"
+        act$lab, " generated an S3 error and you are testing the error message.\n",
+        "* The error has class = c(", klass, ")\n",
+        "* You are testing using `regexp`, not `class`, which is fragile.\n",
+        "* Do you want `expect_error(..., class = \"", class(act$cap)[[1]], "\")`?"
       ))
     }
   }