+++ R/compare.R
@@ -5,8 +5,6 @@
 #'
 #' @export
 #' @param x,y Objects to compare
-#' @param tolerance Numerical tolerance: any differences smaller than this
-#'   value will be ignored.
 #' @param ... Additional arguments used to control specifics of comparison
 #' @keywords internal
 compare <- function(x, y, ...) {