+++ R/expect-output.R
@@ -1,13 +1,31 @@
-#' Expectation: does code produce output/message/warning/error?
+#' Expectation: does code produce errors/warnings/messages?
 #'
-#' Use `expect_output()`, `expect_message()` and `expect_warning()` to
-#' match specified outputs. Use `expect_error()` or `expect_condition()` to
-#' match individual errors or conditions. Use `expect_silent()` to assert that
-#' there should be no output of any type.
+#' Use `expect_message()` and `expect_warning()` to match specified outputs.
+#' Use `expect_error()` or `expect_condition()` to match individual errors or
+#' conditions.
 #'
 #' Note that warnings are captured by a custom signal handler: this means
 #' that `options(warn)` has no effect.
 #'
+#' @section Testing `message` vs `class`:
+#' When checking that code generates an error, it's important to check that the
+#' error is the one you expect. There are two ways to do this. The first
+#' way is the simplest: you just provide a `regexp` that match some fragment
+#' of the error message. This is easy, but fragile, because the test will
+#' fail if the error message changes (even if its the same error).
+#'
+#' A more robust way is to test for the class of the error, if it has one.
+#' You can learn more about custom conditions at
+#' <https://adv-r.hadley.nz/conditions.html#custom-conditions>, but in
+#' short, errors are S3 classes and you can generate a custom class and check
+#' for it using `class` instead of `regexp`. Because this is a more reliable
+#' check, you `expect_error()` will warn if the error has a custom class but
+#' you are testing the message. Eliminate the warning by using `class` instead
+#' of `regexp`.
+#'
+#' If you are using `expect_error()` to check that an error message is
+#' formatted in such a way that it makes sense to a human, we now recommend
+#' using [verify_output()] instead.
 #' @inheritParams expect_that
 #' @inheritDotParams expect_match -object -regexp -info -label
 #' @param regexp regular expression to test against.
@@ -25,15 +43,6 @@
 #' @return The first argument, invisibly. If `expect_error()` captures an
 #'   error, that is returned instead of the value.
 #' @examples
-#' # Output --------------------------------------------------------------------
-#' str(mtcars)
-#' expect_output(str(mtcars), "32 obs")
-#' expect_output(str(mtcars), "11 variables")
-#'
-#' # You can use the arguments of grepl to control the matching
-#' expect_output(str(mtcars), "11 VARIABLES", ignore.case = TRUE)
-#' expect_output(str(mtcars), "$ mpg", fixed = TRUE)
-#'
 #' # Messages ------------------------------------------------------------------
 #'
 #' f <- function(x) {
@@ -68,52 +77,9 @@
 #'
 #' # You can use the arguments of grepl to control the matching
 #' expect_error(f(), "my error!", ignore.case = TRUE)
-#'
-#' # Silent --------------------------------------------------------------------
-#' expect_silent("123")
-#'
-#' f <- function() {
-#'   message("Hi!")
-#'   warning("Hey!!")
-#'   print("OY!!!")
-#' }
-#' \dontrun{
-#' expect_silent(f())
-#' }
 
 #' @name output-expectations
 NULL
-
-#' @export
-#' @rdname output-expectations
-#' @inheritParams capture_output
-expect_output <- function(object,
-                          regexp = NULL,
-                          ...,
-                          info = NULL,
-                          label = NULL,
-                          width = 80
-                          ) {
-  act <- quasi_capture(enquo(object), label, capture_output, width = width)
-
-  if (identical(regexp, NA)) {
-    expect(
-      identical(act$cap, ""),
-      sprintf("%s produced output.\n%s", act$lab, encodeString(act$cap)),
-      info = info
-    )
-  } else if (is.null(regexp) || identical(act$cap, "")) {
-    expect(
-      !identical(act$cap, ""),
-      sprintf("%s produced no output", act$lab),
-      info = info
-    )
-  } else {
-    expect_match(act$cap, enc2native(regexp), ..., info = info, label = act$lab)
-  }
-
-  invisible(act$val)
-}
 
 
 #' @export
@@ -131,10 +97,12 @@ expect_error <- function(object,
 
   if (!is.null(act$cap)) {
     if (!simple_error(act$cap) && is.null(class) && !is.null(regexp)) {
-      klass <- paste0(class(act$cap), collapse = "/")
+      klass <- paste0(encodeString(class(act$cap), quote = '"'), collapse = ", ")
       warn(paste0(
-        act$lab, " generated a condition with class ", klass, ".\n",
-        "It is less fragile to test custom conditions with `class`"
+        act$lab, " generated an S3 error and you are testing the error message.\n",
+        "* The error has class = c(", klass, ")\n",
+        "* You are testing using `regexp`, not `class`, which is fragile.\n",
+        "* Do you want `expect_error(..., class = \"", class(act$cap)[[1]], "\")`?"
       ))
     }
   }
@@ -197,8 +165,23 @@ expect_warning <- function(object,
   invisible(act$val)
 }
 
+#' Expectation: is the code silent?
+#'
+#' Checks that the code produces no output, messages, or warnings.
+#'
+#' @inheritParams expect_error
 #' @export
-#' @rdname output-expectations
+#' @examples
+#' expect_silent("123")
+#'
+#' f <- function() {
+#'   message("Hi!")
+#'   warning("Hey!!")
+#'   print("OY!!!")
+#' }
+#' \dontrun{
+#' expect_silent(f())
+#' }
 expect_silent <- function(object) {
   act <- quasi_capture(enquo(object), NULL, evaluate_promise)
 
@@ -214,6 +197,51 @@ expect_silent <- function(object) {
   )
 
   invisible(act$cap$result)
+}
+
+#' Expectation: does code print output to the console?
+#'
+#' Test for output produced by `print()` or `cat()`. This is best used for
+#' very simple output; for more complex cases use [verify_output()].
+#'
+#' @export
+#' @inheritParams expect_error
+#' @inheritParams capture_output
+#' @inheritDotParams expect_match -object -regexp -info -label
+#' @examples
+#' str(mtcars)
+#' expect_output(str(mtcars), "32 obs")
+#' expect_output(str(mtcars), "11 variables")
+#'
+#' # You can use the arguments of grepl to control the matching
+#' expect_output(str(mtcars), "11 VARIABLES", ignore.case = TRUE)
+#' expect_output(str(mtcars), "$ mpg", fixed = TRUE)
+expect_output <- function(object,
+                          regexp = NULL,
+                          ...,
+                          info = NULL,
+                          label = NULL,
+                          width = 80
+                          ) {
+  act <- quasi_capture(enquo(object), label, capture_output, width = width)
+
+  if (identical(regexp, NA)) {
+    expect(
+      identical(act$cap, ""),
+      sprintf("%s produced output.\n%s", act$lab, encodeString(act$cap)),
+      info = info
+    )
+  } else if (is.null(regexp) || identical(act$cap, "")) {
+    expect(
+      !identical(act$cap, ""),
+      sprintf("%s produced no output", act$lab),
+      info = info
+    )
+  } else {
+    expect_match(act$cap, enc2native(regexp), ..., info = info, label = act$lab)
+  }
+
+  invisible(act$val)
 }