+++ tests/testthat/test-expect-condition.R
@@ -133,4 +133,10 @@ test_that("is_informative_error returns TRUE for basic errors", {
   expect_false(is_informative(abort("!", .subclass = "Rcpp::exception")))
 
   expect_true(is_informative(abort("!", .subclass = "error_custom")))
+
+  with_bindings(
+    .env = global_env(),
+    is_informative_error.error_custom = function(...) FALSE,
+    expect_false(is_informative(abort("!", .subclass = "error_custom")))
+  )
 })