+++ NEWS.md
@@ -3,6 +3,11 @@
 
 * `verify_output()` no longer uses quasiquotation, which fixes issues
   when verifying the output of tidy eval functions (#945).
+* `verify_output()` gains a `unicode` parameter to turn on or off the
+  use of Unicode characters by the cli package. It is disabled by
+  default to prevent the tests from failing on platforms like Windows
+  that don't support UTF-8 (which could be your contributors or your
+  CI machines).
 
 * `verify_output()` now correctly handles multi-line condition
   messages.