+++ NEWS.md
@@ -1,4 +1,4 @@
-# testthat (development version)
+# testthat 2.3.0
 
 ## Conditions